backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.54.0
pydantic>=2.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.27.0
